Sniping and Spotting QoL (#10896)
# About the pull request

Every time a sniper rifle or a spotter designator are stored in
inventory or dropped on the ground, their abilities are deactivated and
have to be toggled back on which is quite annoying.

With this PR "Aimed Shot" and "Spot Target" are activated when their
respective items are equipped in hands.

Works with all "Aimed Shot" capable sniper rifles.

# Explain why it's good for the game

The good:
QoL. Smoother sniping and spotting experience.

The bad:
If by chance you have another item which grants you a middle/right click
ability, picking up spotter designator or sniper rifle will override the
active ability.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/ca94cd0f-9c86-4fe7-b71d-0e950922461c

</details>


# Changelog

:cl: Labeo0
qol: auto-toggle "Aimed Shot"/"Spot Target" with sniper rifle/spotter
designator in hands
/:cl: