Hivelord Whisperer Construction Buff (#7897)
# About the pull request

This PR normalizes the unreasonable restrictions of the Hivelord
Whisperer caste, making it playable. We remove the annoyances that were
broken half a year ago due to attempts to nerf the queen's remote
building.

# Explain why it's good for the game

I am a Hivelord main. It is my favorite caste, and I really enjoyed
playing as Whisperer. But it was nerfed in PR:
https://github.com/cmss13-devs/cmss13/pull/7376



<details>
<summary>Big Scary Text</summary>

### So what do we have:
- Inability to build acid towers (That's fine)
- Your plasma recovery is getting worse.
- You are weak. New AP bullets kill you in 0.5 seconds, you don't even
have time to move away.
- Reduced speed with the inability to accelerate (You are at a distance,
so ok.)
- Reduced speed when you look at the distance (Also fair, you are like a
boiler!)
- You can't build if there is no resin nearby (fair).
- You can build only 1 wall (Like a drone! Wow!)
Very slow construction (3 times slower than a drone)
- Construction of regular walls (Like a drone, yeah)
- You need to wait 3 times longer to place the next wall (Why?)
- If someone or you yourself canceled your construction - then you wait
x*3 cooldown time before you place the next building (oh, did you
accidentally move to the side or the marine removed your blueprint? Wait
10 seconds before you place it, not counting the seconds you wait during
installation and get as an additional cooldown)

The result? A lot of restrictions. What do we get for them? The ability
to place 1 wall every 3-10 seconds at a distance. And it is desirable
that no one interferes with you and you are not in combat. But if so,
then what is the point of taking this caste, which is contraindicated to
be where it would be most useful? A regular drone and a Hivelord will
cope with this more effectively both on the front line and in the rear.

We have been playing with this PR for almost half a year now, which was
aimed more at nerfing the queen than Hivelord Whisperer. I talked to
other maintainers and we came to the conclusion that it needs to be
normalized again, so that it would be interesting to play again, because
now it is trash that annoys and irritates (not the marines, but you, who
plays it).

## My proposal.
### As before:
- We do not touch the Queen in the balance of the nerf of remote
buildings, as the author initially assumed.
- We leave restrictions aimed at movement speed
- You still can't build a tower
- You still only build regular walls
- You also do not have enough plasma for long-term construction, you
need to cooperate.
- We can still build only 1 wall/resin at a time. Until it is built, we
cannot start building a new one.
- No increase in movement speed on resin, unlike the default.

</details>

### Changes:
- No penalty in the form of an 8-10 second cooldown after canceling
construction. There is 2 seconds cooldown again as before.
- Whisperer now AGAIN builds slower than a drone, but faster than a
queen.
- Minor refactor of variables.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/user-attachments/assets/e3d2cf48-bb01-416a-8db0-bcfd33a0d62e
(Old video, now it have penalty like 2 seconds when cancel) 


</details>


# Changelog
:cl:
balance: Hivelord Whisperer build cooldown: 2.5 => 1.5
balance: Hivelord Whisperer building penalty has been reduced: 8 => 2
/:cl: