Hull Windows Overlooking ERT Dock (#10983)
# About the pull request

Makes the windows overlooking the ERT landing docks hull windows.

# Explain why it's good for the game

This is to prevent situations where Marines take position above the ERT
dock and can then attack anyone landing in the ERT dock. Due to a lack
of roof and the general lack of cover this can result in a lot of ERT
casualties.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: The windows overlooking the emergency response team landing
docks on the USS Almayer are now hull windows. This will protect the
shuttle and its occupants from being fired on through the roof of the
shuttle.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>